Share compiled relation plans and isolate mutation execution
