+++ lib/core/relationPlan.ts
@@ -1,17 +1,25 @@
 import type { QueryAST } from './queryBuilder.js'
-import { planRelation } from './queryClassification.js'
-import { relationKey } from './relationalAssembly.js'
+import {
+  explainQueryNode,
+  planRootPagination,
+  rootAllPages,
+  type ExplainNode,
+  type QueryNodeClass,
+} from './queryClassification.js'
 import { resolveServicePath, type RelationshipDef, type Schema } from './schema.js'
 import type { FindDescriptor, FindQueryConfig } from './queryTypes.js'
 
 type RelationValue = string | number
 
 interface RelationQueryPlan {
+  service: string
+  query: Record<string, unknown>
   descriptor(value: RelationValue | { $in: RelationValue[] }): FindDescriptor
   config: FindQueryConfig
 }
 
 interface PlannedRelation {
+  name: string
   key: string
   definition: RelationshipDef
   children: RelationPlan[]
@@ -32,11 +40,18 @@ function queryPlan(
   config: FindQueryConfig,
 ): RelationQueryPlan {
   const serviceName = resolveServicePath(schema, service)
+  const bind = (value: RelationValue | { $in: RelationValue[] }) => ({
+    ...before,
+    [field]: value,
+    ...after,
+  })
   return {
+    service,
+    query: bind({ $in: [] }),
     descriptor: value => ({
       serviceName,
       method: 'find',
-      params: { query: { ...before, [field]: value, ...after } },
+      params: { query: bind(value) },
     }),
     config,
   }
@@ -50,11 +65,18 @@ export function compileRelations(
   parentKey: string | null = null,
 ): RelationPlan[] {
   return Object.entries(ast.related).map(([name, child]) => {
-    const key = relationKey(parentKey, name)
+    const key = parentKey ? `${parentKey}.${name}` : name
     const definition = schema.relationships?.[ast.service]?.[name]
     if (!definition) return { kind: 'missing', key, name, service: ast.service }
-    const { strategy, allPages } = planRelation(definition, child.query)
-    const hasSort = '$sort' in child.query || '$sort' in (definition.query ?? {})
+    const query = { ...child.query, ...definition.query }
+    const windowed = '$limit' in query || '$skip' in query
+    const strategy = definition.via
+      ? 'junction'
+      : windowed && definition.cardinality === 'many'
+        ? 'perParent'
+        : 'fanIn'
+    const allPages = !windowed
+    const hasSort = '$sort' in query
     const sort =
       strategy !== 'perParent' &&
       !hasSort &&
@@ -62,6 +84,7 @@ export function compileRelations(
         ? { $sort: { [definition.destField]: 1 } }
         : {}
     const common = {
+      name,
       key,
       definition,
       children: compileRelations(child, schema, realtime, key),
@@ -93,4 +116,85 @@ export function compileRelations(
     }
     return { ...common, kind: strategy === 'perParent' ? 'perParent' : 'fanIn' }
   })
+}
+
+/** Explain the same resolved queries used by subscriptions and assembly. */
+export function explainQuery(
+  ast: QueryAST,
+  schema: Schema,
+  localOperatorsFor: (serviceName: string) => ReadonlySet<string>,
+  hasNativePagination: (serviceName: string) => boolean,
+): ExplainNode[] {
+  const snapshot = Boolean(ast.snapshot)
+  const pagination =
+    ast.kind === 'paginate'
+      ? planRootPagination(hasNativePagination(ast.service), Boolean(ast.server))
+      : null
+  const root = explainQueryNode(ast.query, {
+    server: pagination?.server ?? ast.server,
+    ...(pagination ? { serverReasons: pagination.serverReasons } : {}),
+    allPages: rootAllPages(ast.kind),
+    localOperators: localOperatorsFor(ast.service),
+    snapshot,
+    paginatedRoot: pagination?.kind === 'offset',
+  })
+  const nodes: ExplainNode[] = [
+    {
+      path: '(root)',
+      service: ast.service,
+      kind: ast.kind,
+      class: root.class,
+      reasons: root.reasons,
+      realtime: realtime(root.class),
+    },
+  ]
+  visit(compileRelations(ast, schema, snapshot ? 'disabled' : 'merge'))
+  return nodes
+
+  function realtime(classification: QueryNodeClass): ExplainNode['realtime'] {
+    return snapshot ? 'manual' : classification === 'local-exact' ? 'merge' : 'refetch'
+  }
+
+  function explain(plan: RelationQueryPlan) {
+    return explainQueryNode(plan.query, {
+      server: plan.config.server,
+      allPages: plan.config.allPages,
+      localOperators: localOperatorsFor(plan.service),
+    })
+  }
+
+  function visit(plans: RelationPlan[]): void {
+    for (const plan of plans) {
+      if (plan.kind === 'missing') continue
+      if (plan.kind === 'junction') {
+        const explained = explain(plan.junction)
+        nodes.push({
+          path: `${plan.key}#junction`,
+          service: plan.junction.service,
+          kind: 'find',
+          role: 'junction',
+          class: explained.class,
+          reasons: explained.reasons,
+          realtime: realtime(explained.class),
+        })
+      }
+      const explained = explain(plan.destination)
+      if (plan.kind === 'perParent') {
+        explained.reasons.push({
+          code: 'window-filter',
+          detail: 'per-parent window — one query per parent',
+        })
+      }
+      nodes.push({
+        path: plan.key,
+        service: plan.destination.service,
+        kind: 'find',
+        class: explained.class,
+        reasons: explained.reasons,
+        realtime: realtime(explained.class),
+        ...(plan.kind === 'junction' ? { via: plan.junction.service } : {}),
+      })
+      visit(plan.children)
+    }
+  }
 }
